% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregation_to_daily}
\alias{aggregation_to_daily}
\title{Aggregate Subdaily Rasters to Daily Values}
\usage{
aggregation_to_daily(subdaily_raster, fun = "none")
}
\arguments{
\item{subdaily_raster}{A \code{SpatRaster} object containing subdaily data. Each layer should correspond
to a time step, and layer names should be formatted as "YYYY-MM-DD HH:MM" to facilitate date extraction.}

\item{fun}{A string specifying the aggregation function. Options are "mean" or "sum".}
}
\value{
A multi-layer \code{SpatRaster} object containing aggregated daily values, with layers named
in the format "YYYY-MM-DD".
}
\description{
This function takes a multi-layer \code{SpatRaster} object containing subdaily data (e.g., hourly)
and aggregates the values to a daily level. The layers are grouped by their corresponding dates,
and an aggregation function (mean or sum) is applied to each group of layers to produce a single
daily layer. The resulting layers are named according to the corresponding dates.
}
\details{
Aggregate Subdaily Rasters to Daily Values
}
\examples{
  # Assuming 'your_subdaily_raster' is a SpatRaster containing hourly data:
  daily_aggregated_raster <- aggregation_to_daily(your_subdaily_raster, fun = "mean")

  # Alternatively, to compute daily sums instead:
  daily_sum_raster <- aggregation_to_daily(your_subdaily_raster, fun = "sum")
}
