% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{app}
\alias{app}
\alias{run}
\title{Launch the HEAT application}
\usage{
app(launch_browser = TRUE)

run()
}
\arguments{
\item{launch_browser}{One of \code{TRUE} or \code{FALSE} specifying if the application
launches in the browser or rstudio viewer, defaults to \code{TRUE}.}
}
\description{
Use \code{run()} to launch the application.
}
