% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{assign_weighting_periods}
\alias{assign_weighting_periods}
\title{Split Raster by Weighting Periods}
\usage{
assign_weighting_periods(raster, period_defs, verbose = 1)
}
\arguments{
\item{:}{\itemize{
\item raster: A cropped SpatRaster with layernames corresponding to the date e.g. name of layer 1: 2017-01-01
\itemize{
\item period_defs: A data.frame containing weighting period definitions with the following columns:
- start_date: The start date of the period.
- end_date: The end date of the period.
}
}}
}
\value{
:
\itemize{
\item A named list of SpatRaster objects, each corresponding to one period.
}
}
\description{
This function splits an already cropped multi-layer SpatRaster into a list of raster stacks,
one for each weighting period defined in the provided period definitions. The period definitions
should be a data.frame with 'start_date' and 'end_date' columns (in Date format). The function
returns a named list of raster stacks, where each element corresponds to a period (e.g., "period_1").
}
\examples{
\dontrun{
  # Assuming 'cropped_raster' is a cropped SpatRaster with time metadata and
  # 'period_defs' is a data.frame with start_date and end_date columns:
  periods_list <- assign_weighting_periods(cropped_raster, period_defs)
}
}
