% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data.R
\name{buffer_extent}
\alias{buffer_extent}
\title{Buffer Spatial Extent Based on Raster Resolution}
\usage{
buffer_extent(raster, spatial_data, buffer_factor = 1)
}
\arguments{
\item{raster}{A SpatRaster object from which the resolution is extracted.}

\item{spatial_data}{An sf or SpatVector object containing points or polygons.}

\item{buffer_factor}{Numeric factor by which to multiply the raster resolution when buffering
the spatial extent. Default is 1. Use 0 for no buffer.}
}
\value{
A buffered extent (terra::ext object) computed as:
xmin = spatial_extent$xmin - buffer_factor * x_res,
xmax = spatial_extent$xmax + buffer_factor * x_res,
ymin = spatial_extent$ymin - buffer_factor * y_res,
ymax = spatial_extent$ymax + buffer_factor * y_res.
}
\description{
This function computes a buffered extent around the input spatial data (points or polygons)
using the raster's resolution. The extent is expanded by a factor specified by the user
(buffer_factor). This is useful for ensuring that point locations fall within raster cells
when using terra::window(), or for adding a buffer around polygon boundaries.
}
\examples{
\dontrun{
  # Buffer extent by one cell resolution (useful for points)
  buffered_extent <- buffer_extent(environmental_raster, points, buffer_factor = 1)
  
  # No buffer (useful for polygons)
  unbuffered_extent <- buffer_extent(environmental_raster, polygons, buffer_factor = 0)
}
}
