% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data.R
\name{buffer_polygons}
\alias{buffer_polygons}
\title{Buffer Polygons Based on Raster Resolution}
\usage{
buffer_polygons(raster, polygons, buffer_factor)
}
\arguments{
\item{:}{\itemize{
\item raster: (SpatRaster) A raster object from which the resolution is extracted.
\itemize{
\item polygons: (sf or SpatVector) A spatial object containing the polygons.
\item buffer_factor: (numeric) A factor by which to multiply the raster resolution when buffering
the polygons' extent.
}
}}
}
\value{
:
\itemize{
\item A buffered extent (terra::ext object) computed as:
xmin = polygons_extent$xmin - buffer_factor * x_res,
xmax = polygons_extent$xmax + buffer_factor * x_res,
ymin = polygons_extent$ymin - buffer_factor * y_res,
ymax = polygons_extent$ymax + buffer_factor * y_res.
}
}
\description{
This function computes a buffered extent around the input polygons using the raster's resolution.
The extent is expanded by a factor specified by the user (buffer_factor). The buffered extent is
returned as a terra::ext object.
}
\examples{
  buffered_extent <- buffer_polygons(environmental_raster, polygons, buffer_factor = 1)
}
