% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uval.R
\name{calculateUval}
\alias{calculateUval}
\title{calculateUval}
\usage{
calculateUval(Layers, k.t_e = -10, k.t_i = 20, k.h_e = 25, k.h_i = 8,
  k.rf_e = 80, k.rf_i = 50, k.p_e = 1168, k.p_i = 321,
  plottemp = FALSE, plotpress = FALSE, writetab = FALSE, name = "")
}
\arguments{
\item{Layers,}{building elements layers}

\item{k.t_e}{(optional, default = -10) External temperature}

\item{k.t_i}{(optional, default = 20) Internal temperature}

\item{k.h_e}{(optional, default = 25)}

\item{k.h_i}{(optional, default = 8) Internal heat transfer coefficients [m2K/W]}

\item{k.rf_e}{(optional, default = 80)}

\item{k.rf_i}{(optional, default  = 50) Internal air moisture}

\item{k.p_e}{(optional, default = 1168)}

\item{k.p_i}{(optional, default = 321) Internal pressure [Pa]}

\item{plottemp}{(optional, default = FALSE)}

\item{plotpress}{(optional, default = FALSE)}

\item{writetab}{(optional, default = FALSE)}

\item{name}{(optional = "")}
}
\value{
Uval (optional = component U-value)
}
\description{
U-value calculator
}
\author{
M. Estebna Munoz H.
}

