% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_main.R
\name{check_na_in_date_sequences}
\alias{check_na_in_date_sequences}
\title{Function to check for NAs in date sequences for each row}
\usage{
check_na_in_date_sequences(
  windows_data,
  polygon_id,
  main_dt,
  metad_cols,
  climate_dates
)
}
\arguments{
\item{windows_data:}{A data.table object with data corresponding to a given window sequence.}

\item{polygon_id:}{A character string indicating the ID of the geometry.}

\item{main_dt:}{The data.table object to subset the sequence data from.}

\item{metad_cols:}{A vector with the name of the metadata columns in the parquet file.}

\item{climate_dates:}{A vector with the parquet file date columns}
}
\value{
: named list if there are any NA values found for a given polygon and date sequences. Otherwise, returns FALSE (no NAs found)
}
\description{
This function checks if there are any NAs in the data corresponding to a given date sequence
}
