% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{check_packages}
\alias{check_packages}
\title{Checks if all required packages are installed.}
\usage{
check_packages()
}
\value{
\itemize{
\item NULL
}
}
\description{
This functions checks if all packages required by the shock functions are installed and installs them if not.
}
