% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{check_spatial_agg_args}
\alias{check_spatial_agg_args}
\title{Check Spatial Aggregation Arguments}
\usage{
check_spatial_agg_args(args)
}
\arguments{
\item{args}{A list of spatial aggregation arguments.}
}
\value{
A (possibly modified) list of spatial aggregation arguments.
}
\description{
This function verifies that the supplied spatial aggregation arguments are valid.
If the required arguments are not provided, it sets them to default values:
\itemize{
\item fun: "weighted_mean"
\item stack_apply: FALSE
Additionally, if an appended columns argument ("append_cols") is present, it is removed,
and a message is printed indicating that these columns will be added manually later.
}
}
\examples{
spatial_agg_args <- list(
  append_cols = c("col1", "col2")
)
updated_args <- check_spatial_agg_args(spatial_agg_args)
# updated_args now contains fun = "weighted_mean", stack_apply = FALSE, and append_cols = NULL.
}
