% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{check_trans}
\alias{check_trans}
\title{Check Transformation Function Arguments}
\usage{
check_trans(fun, args)
}
\arguments{
\item{:}{\itemize{
\item fun: (function) The transformation function to be used (e.g., stats::poly, splines::bs).
\itemize{
\item args: (list) A list of arguments intended for the transformation function.
}
}}
}
\value{
:
\itemize{
\item A (possibly modified) list of arguments, ensuring that required conditions (such as raw = TRUE for poly)
are met.
}
}
\description{
This function verifies that the supplied transformation function is valid and that its
arguments meet the required conditions for processing. In particular, it checks that:
- The provided object is a function.
- The function includes an argument named "x".
- If the function is stats::poly (used for polynomial transformations), it forces the
argument 'raw' to TRUE.
}
\examples{
  # Example for a polynomial transformation:
  trans_args <- list(degree = 3)
  updated_args <- check_trans(stats::poly, trans_args)
  # updated_args now contains degree = 3 and raw = TRUE.
}
