% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{create_dirs}
\alias{create_dirs}
\title{Create intermediate directories if they do not already exist}
\usage{
create_dirs(base_path, subfolders)
}
\arguments{
\item{base_path}{A character string indicating the base folder path}

\item{subfolders}{A vector with subfolders to create inside the baseline path.}
}
\value{
Creates, if necessary, the baseline folders and its respective subfolders
}
\description{
Checks if a directory exists and creates it if not
}
