% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{create_pairs}
\alias{create_pairs}
\title{Return a dataframe with a column where a unique ID in the form polygonid_date is included}
\usage{
create_pairs(pol_date_pairs, geom_id_col = "geom_id", date_id_col = "date")
}
\arguments{
\item{pol_date_pairs}{dataframe with unique polygon date pairs}

\item{geom_id_col}{character string with the name of the column that identifies geometries}

\item{date_id_col}{character string with the name of the column that identifies dates}
}
\value{
a dataframe equal to the original plus the polygonid_date column
}
\description{
Return a dataframe with a column where a unique ID in the form polygonid_date is included
}
