% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createADJforHT.R
\name{createADJforHT}
\alias{createADJforHT}
\title{createADJforHT}
\usage{
createADJforHT(data, vari, HWvari, maxlag, adj = 1, piecewise = TRUE)
}
\arguments{
\item{data}{dataset to be used.}

\item{vari}{the name of temperature variable to be used in the dataset (e.g., t_mean; not "t_mean").}

\item{HWvari}{the name of the binary variable to indicate the presence of a heatwave (e.g., HW; not "HW")}

\item{maxlag}{the maximum lag period to adjust for temperature}

\item{adj}{if=1 (default), variables for adjustment for temperature including lag effects of heatwaves. if=2, variables for adjustment for temperature excluding lag effects of heatwaves.}

\item{piecewise}{if TRUE (default), piecewise temperature variables for adjustment; if FALSE, non-piecewise temperature variables for adjustment}
}
\description{
This function creates variables for adjustment for confounding by temperature.
}
\examples{
createADJforHT(data=newdata,vari=t_mean,HWvari=HW,maxlag=2,adj=1)
}
