% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFHW.R
\name{createFHW}
\alias{createFHW}
\title{createFHW}
\usage{
createFHW(data, vari, HWvari, maxlag, hw_th, hw_dr, LEHRNH = FALSE)
}
\arguments{
\item{data}{dataset to be used.}

\item{vari}{the name of temperature variable to be used in the dataset (e.g., t_mean; not "t_mean").}

\item{HWvari}{the name of the binary variable to indicate the presence of a heatwave (e.g., HW; not "HW")}

\item{hw_th}{the intensity threshold for the definition of a heatwave (e.g., temperature exceeding the 99th percentile of temperature distribution)}

\item{hw_dr}{the duration threshold for the definition of a heatwave (e.g., at least two consecutive days).}

\item{LEHRNH}{Lag Effects on the Health Risk on the Next Heatwave (LEHRNH). If TRUE, the lag effects of the previous heatwave on the health risk on the next heatwave (if they exist) are subsumed into the effect of the lag0 heatwave indicator variable. Default is FALSE}
}
\description{
This function create lag variables of a binary heatwave indicator variable.
}
\examples{
createFHW(data=data,vari=t_mean,HWvari=HW,maxlag=1,hw_th=quantile(data$t_mean,0.99),hw_dr=2,LEHRNH=FALSE)
}
