% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHW.R
\name{createHW}
\alias{createHW}
\title{createHW}
\usage{
createHW(data, vari, hw_th, hw_dr)
}
\arguments{
\item{data}{dataset to be used.}

\item{vari}{the name of temperature variable to be used in the dataset (e.g., t_mean; not "t_mean").}

\item{hw_th}{the intensity threshold for the definition of a heatwave (e.g., temperature exceeding the 99th percentile of temperature distribution)}

\item{hw_dr}{the duration threshold for the definition of a heatwave (e.g., at least two consecutive days).}
}
\description{
This function create a binary variable to indicate the presense of a heatwave
}
\examples{
createHW(data=data,vari=t_mean,hw_th=quantile(data$t_mean,0.99),hw_dr=2)
}
