% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createV.R
\name{createV}
\alias{createV}
\title{createV}
\usage{
createV(data, vari, maxlag, hw_th, hw_dr)
}
\arguments{
\item{data}{dataset to be used.}

\item{vari}{the name of temperature variable to be used in the dataset (e.g., t_mean; not "t_mean").}

\item{maxlag}{the maximum lag period.}

\item{hw_th}{the intensity threshold for the definition of a heatwave (e.g., temperature exceeding the 99th percentile of temperature distribution).}

\item{hw_dr}{the duration threshold for the definition of a heatwave (e.g., at least two consecutive days).}
}
\description{
This function create vector V to adjust for temperature
}
\examples{
createV(data=data,vari=t_mean,maxlag=1,hw_th=quantile(data$t_mean,0.99),hw_dr=2)
}
