% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_main.R
\name{daily_shocks}
\alias{daily_shocks}
\title{Function to estimate daily shocks}
\usage{
daily_shocks(
  current_ds,
  dates_mat,
  bin,
  return_measure,
  target_df,
  climate_dates,
  hist_lags
)
}
\arguments{
\item{current_ds:}{A vector of contemporary sequence of dates}

\item{dates_mat:}{A matrix where each row corresponds to a historical sequence of dates}

\item{bin:}{The bin id within the bin structure.}

\item{return_measure:}{A character indicating which climate measure to return. Must be: "contemp_avg", "contemp_sd", "historic_avg", "historic_sd")}

\item{target_df:}{A data.table object to perform the calculations on}

\item{climate_dates:}{A vector with the parquet file climate dates. Inherited.}

\item{hist_lags:}{An integer indicating the number of lags to build the historical baseline. Inherited.}
}
\value{
the value of the indicated historical measure
}
\description{
This function estimates climate shocks when the product´s resolution is daily.
}
