% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_main.R
\name{get_data}
\alias{get_data}
\title{Function to retrieve the data from the parquet file}
\usage{
get_data(
  group,
  pol_date_pairs,
  geom_id_col,
  conditions_list,
  start,
  window,
  time_step,
  align,
  hist_lags,
  window_spec,
  start_date,
  stop_date,
  climate_dates,
  int_threshold,
  error_log_path
)
}
\arguments{
\item{group}{An integer indicating the group to which a polygon belongs. This number is assigned by the shocks_wrapper function.}

\item{pol_date_pairs}{A dataframe with unique geometry id and date pairs and a group id. The geometry IDs must correspond to those
in the parquet file with climate measures. Dates must be in date format (eg: "YYYY-MM-DD"). Inherited from shocks_wrapper.}

\item{conditions_list}{A named list of conditions to query the parquet file with climate measures. Inherited from shocks_wrapper.}

\item{window}{An integer indicating the length of the data sequence. Inherited from shocks_wrapper.}

\item{time_step}{A character indicating the product´s resolution}

\item{climate_dates}{A chector with the parquet file date columns.}

\item{int_threshold}{A float number indicating the minimum threshold that date sequences must intersect with the date columns in the parquet file to be considered valid. Inherited from shocks_wrapper.}

\item{geom_id_col:}{a character string indicating the name of the column in the parquet file that identifies the geometries}

\item{start:}{An integer indicating the offset in days with respect to the date polygon date. Inherited from shocks_wrapper.}

\item{align:}{A character string indicating whether the sequence is left- or right-aligned with respect to the polygon date. Inherited from shocks_wrapper.
Left-alignment means that the sequence contains dates AFTER the polygon date. Right-alignment means that the sequence includes dates BEFORE the polygon date. Inherited from shocks_wrapper.}

\item{hist_lags:}{An integer indicating the number of years to construct the historical baseline. Required if window_spec is "dynamic" or "both". Inherited from shocks_wrapper.}

\item{window_spec:}{A character string indicating whether the historical baseline is dynamic, fixed or both. Inherited from shocks_wrapper.}

\item{start_date:}{A character string in date format ("YYYY-MM-DD") indicating the start date of the historical window. Only required if window_spec= "dynamic" or window_spec="both". Inherited from shocks_wrapper.}

\item{stop_date:}{A character string in date format ("YYYY-MM-DD") indicating the end date of the historical window. Only required if window_spec= "dynamic" or window_spec="both". Inherited from shocks_wrapper.}

\item{error_log_path:}{A character string with the path to the folder where the error logs are stored. Inherited from shocks_wrapper.}
}
\value{
a data.table object with selected date columns (as defined in the windows_seq function) and polygons.
}
\description{
Function to retrieve the data from the parquet file
}
