% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{get_period_boundaries}
\alias{get_period_boundaries}
\title{Get the boundary dates for the weighting periods based on the layer names of the secondary weight raster.}
\usage{
get_period_boundaries(sec_weight_raster, boundary_dates)
}
\arguments{
\item{sec_weight_raster}{A SpatRaster object containing secondary weight metrics.}

\item{boundary_dates}{A numeric vector of two dates defining the date range (start and end).}
}
\value{
A list with two elements:
\itemize{
\item A vector of first dates assigned based on the closest secondary weight.
\item A vector of indices corresponding to the secondary weight layers used for each assigned date.
}
}
\description{
This function takes a secondary weight raster and a vector of boundary dates to
generate a sequence of dates. It then determines the temporal resolution of the
raster, assigning the relevant dates to the appropriate year, month, or day
accordingly, and returns the first date in each sequence along with the indices
of the corresponding secondary weight raster layers.
}
