% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_trans_var_cols}
\alias{get_trans_var_cols}
\title{Get Transformation Variable Columns}
\usage{
get_trans_var_cols(df)
}
\arguments{
\item{df}{A data frame to search for transformation variable columns}
}
\value{
A character vector of transformation variable column names
}
\description{
Identifies transformation variable columns in a data frame. Throws error if multiple types found.
Warns if none found.
}
\examples{
\dontrun{
  trans_cols <- get_trans_var_cols(my_data)
}
}
