% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data.R
\name{get_valid_raster_files}
\alias{get_valid_raster_files}
\title{Get Valid Raster Files}
\usage{
get_valid_raster_files(path, pattern = "\\\\.tif$|\\\\.nc$")
}
\arguments{
\item{:}{\itemize{
\item path: A character string specifying the path to the folder containing raster files.
\itemize{
\item pattern: A character string specifying the file pattern to match (default is "\\.tif$|\\.nc$").
}
}}
}
\value{
:
\itemize{
\item A character vector containing paths to the valid raster files that can be read.
}
}
\description{
This function searches for raster files with specific extensions in a given folder,
checking whether they can be read into R. It returns a list of accessible raster files
while printing warnings for any files that cannot be read.
}
\examples{
  valid_rasters <- get_valid_raster_files(paths$path_env_rast)
}
