% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_main.R
\name{get_window_estimates_vectorized}
\alias{get_window_estimates_vectorized}
\title{Inner function to estimate climate measures and shocks from data.table objects}
\usage{
get_window_estimates_vectorized(
  polygon_id,
  all_windows_data,
  bin,
  bin_name,
  main_dt,
  pol_date_pairs_set,
  start,
  window,
  time_step,
  align,
  hist_lags,
  window_spec,
  start_date,
  stop_date,
  metad_cols,
  climate_dates
)
}
\arguments{
\item{window}{An integer indicating the length of the data sequence. Inherited from shocks}

\item{time_step}{product´s resolution. Inherited from shocks.}

\item{metad_cols}{A vector of columns names in the parquet file that are used as metadata columns. Inherited from shocks.}

\item{polygon_id:}{A chatacter string with polygon identifier}

\item{all_windows_data:}{lA ist with contemporary and historic date sequences for a given polygon-date pair}

\item{bin:}{A boolean vector with binning structure}

\item{bin_name:}{A character string indicating the bin name}

\item{main_dt:}{A data.table object with climate data for a given polygon-date pair}

\item{pol_date_pairs_set:}{A subset of the polyigon-dates dataframe}

\item{climate_dates:}{A vector with the parquet file date columns}

\item{start:}{An integer indicating the offset in days with respect to the date polygon date. Inherited from shocks.}

\item{hist_lags:}{An integer indicating the number of years to construct the historical baseline. Required if window_spec is "dynamic" or "both". Inherited from shocks.}

\item{align:}{A character string indicating whether the sequence is left- or right-aligned with respect to the polygon date. Inherited from shocks_wrapper.
Left-alignment means that the sequence contains dates AFTER the polygon date. Right-alignment means that the sequence includes dates BEFORE the polygon date. Inherited from shocks.}

\item{window_spec:}{A character string indicating whether the historical baseline is dynamic, fixed or both. Inherited from shocks.}

\item{start_date:}{A character string in date format ("YYYY-MM-DD") indicating the start date of the historical window. Only required if window_spec= "dynamic" or window_spec="both". Inherited from shocks.}

\item{stop_date:}{A character string in date format ("YYYY-MM-DD") indicating the end date of the historical window. Only required if window_spec= "dynamic" or window_spec="both". Inherited from shocks.}
}
\value{
: A data.table object with climate measures and shocks for a given polygon id
}
\description{
Inner function to estimate climate measures and shocks from data.table objects
}
