% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uval.R
\name{getLayers}
\alias{getLayers}
\title{getLayers}
\usage{
getLayers(material, thickness, materialsData = FALSE, verbose = FALSE)
}
\arguments{
\item{material,}{material names}

\item{thickness,}{layer thickness}

\item{materialsData,}{(optional default = False).}
}
\value{
Layers
}
\description{
Function to generate layers based on an external data source
}
\examples{
Materials <- c(
    "Mineralischer Edelputz",
    "Porensinterbeton mit Quarzsand_900",
    "EPS_040.30",
    "Holzwolle-Leichtbauplatten_Heraklith-Platten (Magnesia)_390",
    "Mineralischer Armierungsputz"
    )

Thicknes = c(0.02,0.24,0.04,0.025,0.015)
Layers <- getLayers(Materials, Thicknes)
}
\author{
M. Estebna Munoz H.
}

