% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{getLW}
\alias{getLW}
\title{getLW}
\usage{
getLW(A, P, tl = 0.1)
}
\arguments{
\item{A}{geometry area}

\item{P}{geometry perimeter}

\item{tl}{(optional, default = 0.1) tolerance level for the approximation}
}
\value{
c(l, w)
}
\description{
Simplifies a complex geometry as length and width maintaining the area and
perimeter of the original geometry.
}
\author{
M. Esteban Munoz H.
}

