% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{getSolarRadiation}
\alias{getSolarRadiation}
\title{getSolarRadiation}
\usage{
getSolarRadiation(building_orientation)
}
\arguments{
\item{building_orientation}{orientation of the building in degrees}
}
\value{
Climate.I radiation levels
}
\description{
Computes the solar radiation of the building based on its orientation
}
\author{
M. Estebna Munoz H.
}

