% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{group_polygons_by_date}
\alias{group_polygons_by_date}
\title{Function to groupo polygons by date proximity}
\usage{
group_polygons_by_date(pol_date_pairs, tolerance_days = 7)
}
\arguments{
\item{pol_date_pairs}{A data.frame with polygons and dates.}

\item{tolerance_days}{An integer indicating the maximum separation in time units tolerated for polygons to be grouped together.}
}
\value{
A data.frame with the polygon groupings by date.
}
\description{
This function groups polygons by date proximity, as indicated by the tolerance_days argument
}
