% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heat.R
\name{heatest}
\alias{heatest}
\title{heat}
\source{
DIN V 18599
}
\usage{
heatest(building = FALSE, climate = FALSE, General.Name = "No Name",
  General.PlotOption = FALSE, General.PlotName = "No Name",
  Building.Uwb = FALSE, Building.Windows = FALSE, Building.UvalW = FALSE,
  Building.UvalR = FALSE, Building.UvalWindow = FALSE,
  Building.Dim = c(FALSE, FALSE), Building.h = FALSE,
  Building.AirCRate = FALSE, Building.Ti = FALSE, Building.qi = FALSE,
  Building.Orientation = FALSE, Building.StorageCapacity = FALSE,
  Building.RoofSlope = FALSE, Output.Type = "Year")
}
\arguments{
\item{building}{(optional, default = FALSE) use a user define building
data file.}

\item{climate}{(optional, default = FALSE) use a specific climate file.}

\item{General.Name}{(optional, default = "No Name") if this value is not
specifically define the function will load the values from the building data
file.}

\item{General.PlotOption}{(optional, default = FALSE)}

\item{General.PlotName}{(optional, default="No Name")}

\item{Building.Uwb}{(optional, default = FALSE)}

\item{Building.Windows}{(optional, default = FALSE)}

\item{Building.UvalW}{(optional, default = FALSE)}

\item{Building.UvalR}{(optional, default = FALSE)}

\item{Building.UvalWindow}{(optional, default = FALSE)}

\item{Building.Dim}{(optional, default = c(FALSE,FALSE))}

\item{Building.h}{(optional, default = FALSE)}

\item{Building.AirCRate}{(optional, default = FALSE)}

\item{Building.Ti}{(optional, default = FALSE)}

\item{Building.qi}{(optional, default = FALSE)}

\item{Building.Orientation}{(optional, default = FALSE)}

\item{Building.StorageCapacity}{(optional, default = FALSE)}

\item{Building.RoofSlope}{(optional, default = FALSE)}

\item{Output.Type}{(optional, default = "Year")}
}
\description{
Perform a monthly heat balance and annual energy demand of buildings
}
\author{
M. Esteban Munoz H.
}

