% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{infer_num_trans_var}
\alias{infer_num_trans_var}
\title{Infer Number of Transformation Variables}
\usage{
infer_num_trans_var(trans_type, trans_args)
}
\arguments{
\item{trans_type}{A character string specifying the type of transformation.
Expected values include "bin", "polynomial", "natural_spline", "b_spline", or "none".}

\item{trans_args}{A list of arguments associated with the transformation type, which may
contain relevant parameters such as breaks, degree, knots, etc.}
}
\value{
An integer representing the number of transformation variables that will be created
based on the logic associated with the transformation type.
}
\description{
This function infers the number of transformation variables that will be created
based on the specified transformation type and its associated arguments.
}
\examples{
  num_vars_bin <- infer_num_trans_var(trans_type = "bin", trans_args = list(breaks = c(10, 20, 30)))
  print(num_vars_bin)  # Expected output: 4

  num_vars_none <- infer_num_trans_var(trans_type = "none", trans_args = list())
  print(num_vars_none)  # Expected output: 1

}
