% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_bin_proportion}
\alias{plot_bin_proportion}
\title{Plot Bin Proportion}
\usage{
plot_bin_proportion(data, temporal_resolution, geom_id_col = "geom_id")
}
\description{
Plot Bin Proportion
}
\keyword{internal}
