% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_spatial_averages}
\alias{plot_spatial_averages}
\title{Plot Spatial Averages}
\usage{
plot_spatial_averages(
  data,
  target_geometry,
  title,
  fill_variable,
  fill_label,
  x_label = "Longitude",
  y_label = "Latitude"
)
}
\description{
Plot Spatial Averages
}
\keyword{internal}
