% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_summary_ridge}
\alias{plot_summary_ridge}
\title{Plot Summary Ridge Plot}
\usage{
plot_summary_ridge(
  data,
  temporal_resolution,
  fill_variable,
  environmental_variable_description
)
}
\description{
Plot Summary Ridge Plot
}
\keyword{internal}
