% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot Time Series}
\usage{
plot_time_series(
  temporal_resolution,
  first_df,
  second_df = NULL,
  third_df = NULL,
  fourth_df = NULL,
  value_variable = "environmental_variable",
  value_label = "Environmental Variable",
  first_title = "Dataset 1",
  second_title = "Dataset 2",
  third_title = "Dataset 3",
  fourth_title = "Dataset 4"
)
}
\description{
Plot Time Series
}
\keyword{internal}
