\name{print.threshpt}
\alias{print.threshpt}

\title{
Print a threshpt Object
}
\description{
A default print method for a \code{threshpt} object
}

\usage{
\method{print}{threshpt}(x, ...)
}

\arguments{
  \item{x}{
a fitted \code{threshpt} object produced by \code{threshpt()}
}
  \item{\dots}{
Not used
}
}

\value{
\code{threshpt} returns a optimal threshold and estimates/etandard errors/p-values for the coefficients of a fitted \code{threshpt} object provided by \code{threshpt()}.
}

\author{
Youn-Hee Lim, Il-Sang Ohn, and Ho Kim
}
