% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{process_data}
\alias{process_data}
\title{Process Data for Validation Checks}
\usage{
process_data(
  dataframe,
  input_temporal_resolution,
  output_temporal_resolution = NULL,
  filter_start_date = NULL,
  filter_end_date = NULL,
  filter_poly_ids = NULL,
  average_over_time = FALSE,
  average_over_poly_ids = FALSE,
  value_columns
)
}
\description{
Process Data for Validation Checks
}
\keyword{internal}
