% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{relative_baseline}
\alias{relative_baseline}
\title{Function to create a relative baseline}
\usage{
relative_baseline(d, hist_lags = 10, width = 5, align = "center")
}
\arguments{
\item{d}{date in YYYY-MM-DD format}

\item{hist_lags}{An integer indicating how many year lags are used to build the historical window}

\item{width}{An integer indicating the size of the window sequence}

\item{align}{An integer indicating how the window sequence is aligned with respect to the date}
}
\value{
A vector of dates
}
\description{
This function takes as input a date and creates a relative baseline following the hist_lags, width and align parameters.
}
\examples{
\dontrun{
historic_baseline <- relative_baseline("2000-01-01", hist_lags = 10, width = 10, align = "center")
}
}
