% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{select_trans_fun}
\alias{select_trans_fun}
\title{Select Transformation Function}
\usage{
select_trans_fun(trans_type)
}
\arguments{
\item{:}{\itemize{
\item trans_type: (character) A string specifying the desired transformation type.
Supported values are "polynomial", "natural_spline", "b_spline", and "custom".
}}
}
\value{
:
\itemize{
\item A function object corresponding to the selected transformation. For example, if
trans_type is "polynomial", it returns stats::poly.
}
}
\description{
This function selects and returns the appropriate transformation function based on the
provided transformation type. It supports the following transformation types:
- "polynomial": returns the polynomial transformation function from stats::poly.
- "natural_spline": returns the natural spline transformation function from splines::ns.
- "b_spline": returns the B-spline transformation function from splines::bs.
- "custom": returns NULL and prompts the user to assign a custom transformation function.
}
\examples{
  # Select natural spline transformation
  ns_fun <- select_trans_fun("natural_spline")

  # Select B-spline transformation
  bs_fun <- select_trans_fun("b_spline")
}
