% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_main.R
\name{shocks}
\alias{shocks}
\title{Function to estimate climate shocks}
\usage{
shocks(
  n_groups,
  n_workers,
  pol_date_pairs_grouped,
  conditions_list,
  geom_id_col,
  window,
  start,
  hist_lags,
  align,
  time_step_size,
  window_spec,
  prod_res,
  start_date,
  stop_date,
  met_cols,
  climate_dates,
  int_threshold,
  query,
  output_path,
  error_log_path,
  disjoint = FALSE
)
}
\arguments{
\item{n_groups}{An integer indicating the total number of polygon groups. Estimated by the shocks_wrapper function.}

\item{n_workers}{An integer indicating how many cores are used for parallelization. Inherited from shocks_wrapper.}

\item{pol_date_pairs_grouped}{A dataframe with unique geometry id and date pairs and a group id. The geometry IDs must correspond to those
in the parquet file with climate measures. Dates must be in date format (eg: "YYYY-MM-DD"). Inherited from shocks_wrapper.}

\item{conditions_list}{A named list of conditions to query the parquet file with climate measures. Inherited from shocks_wrapper.}

\item{window}{An integer indicating the length of the data sequence. Inherited from shocks_wrapper.}

\item{met_cols}{A vector of columns names in the parquet file that are used as metadata columns. Inherited from shocks_wrapper.}

\item{int_threshold}{A float number indicating the minimum threshold that date sequences must intersect with the date columns in the parquet file to be considered valid. Inherited from shocks_wrapper.}

\item{geom_id_col:}{a character string indicating the name of the column in the parquet file that identifies the geometries}

\item{start:}{An integer indicating the offset in days with respect to the date polygon date. Inherited from shocks_wrapper.}

\item{hist_lags:}{An integer indicating the number of years to construct the historical baseline. Required if window_spec is "dynamic" or "both". Inherited from shocks_wrapper.}

\item{align:}{A character string indicating whether the sequence is left- or right-aligned with respect to the polygon date. Inherited from shocks_wrapper.
Left-alignment means that the sequence contains dates AFTER the polygon date. Right-alignment means that the sequence includes dates BEFORE the polygon date. Inherited from shocks_wrapper.}

\item{time_step_size:}{An integer indicating the number of days in each temporal bin that divides the date sequence. Inherited from shocks_wrapper.}

\item{disjoint:}{A boolean indicating whether the bins are disjoint or overlapping. Inherited from shocks_wrapper.}

\item{window_spec:}{A character string indicating whether the historical baseline is dynamic, fixed or both. Inherited from shocks_wrapper.}

\item{start_date:}{A character string in date format ("YYYY-MM-DD") indicating the start date of the historical window. Only required if window_spec= "dynamic" or window_spec="both". Inherited from shocks_wrapper.}

\item{stop_date:}{A character string in date format ("YYYY-MM-DD") indicating the end date of the historical window. Only required if window_spec= "dynamic" or window_spec="both". Inherited from shocks_wrapper.}

\item{max_group_size:}{A integer indicating the maximum number of polygons in each processing group. Inherited from shocks_wrapper.}

\item{date_tolerance_days:}{An integer indicating the maximum number of dates that two polygon interviews should be apart to consider them part of two different polygon groups. Inherited from shocks_wrapper.}

\item{prop_cores:}{A float number indicating the proportion of cores to use for parallelization. Inherited from shocks_wrapper.}

\item{query:}{A character string with the name given to the output. Inherited from shocks_wrapper.}

\item{output_path:}{A character string with the path to the folder where the output should be stored. Inherited from shocks_wrapper.}

\item{error_log_path:}{A character string with the path to the folder where the error logs are stored. Inherited from shocks_wrapper.}
}
\value{
: A a list with metadata (i.e., file paths to intermediate parquet files and error logs).
}
\description{
This function estimates the climate shocks as well as the climate measures (i.e., mean, SD) for each polygon-date pair.
Several arguments used in this function are inherited from the shocks_wrapper function.
}
\details{
This function exports intermediate parquet files with climate measures and shocks, as well as any error lists if produced.
The function performs the following operations:
\itemize{
\item Estimates the number of binds within the date sequence according to the bin width. Automatically adjusts for unequal bin sizes.
\item Process in parallel (or sequentially if n_workers = 1) the estimation of shocks for the polygon groups.
\item Loads data from the parquet file corresponding only to the polygons in a specific group and according to the conditions list.
\item Checks for NA in original parquet file and produces error logs indicating which polygons and dates have NAs, if any.
\item Estimates climate measures and shocks for each polygon-date pair, according to the conditions list and the binning structure.
\item Stores temporal intermediate parquet files in the intermediate folder.
\item Returns a list with metadata about the intermediate objects´ and the error logs´ paths.
}
}
