% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{sinusoidal_interpolation}
\alias{sinusoidal_interpolation}
\title{Sinusoidal Interpolation for Hourly Temperature Values}
\usage{
sinusoidal_interpolation(tmin, tmax, hour = 1:24)
}
\arguments{
\item{tmin}{A numeric vector of daily minimum temperatures (tmin), where each value corresponds to a day.}

\item{tmax}{A numeric vector of daily maximum temperatures (tmax), where each value corresponds to a day.}

\item{hour}{A numeric vector representing the hours of the day (1 to 24).}
}
\value{
A numeric vector containing the computed hourly temperatures for each hour across all days represented
in the input tmin and tmax vectors. The length of the returned vector will be 24 times the number of days
represented by the input vectors. The output layer are named in format "YYYY-MM-DD HH:MM", where the date corresponds to the day of the respective tmin and tmax values.
}
\description{
This function computes hourly temperature interpolations from daily minimum and
maximum temperature values using a sinusoidal approach. It calculates interpolated
hourly temperatures for each day represented in the input tmin and tmax vectors.
}
\examples{
  # Assuming tmin and tmax are numeric vectors of length equal to the number of days:
  hourly_temperatures <- sinusoidal_interpolation(tmin = c(15, 16, 14), 
                                                  tmax = c(22, 23, 21), 
                                                  hour = 1:24)
}
