% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{split_large_groups}
\alias{split_large_groups}
\title{Creates groups of polygons to ensure they are all sized max_group_size or smaller}
\usage{
split_large_groups(grouped_pol_pairs, max_polygons_per_group)
}
\arguments{
\item{grouped_pol_pairs}{A data.frame with polygons grouped by date proximity.}

\item{max_polygons_per_group}{maximum number of polygons allowed per group.}
}
\value{
A data.frame with new groupings assigned.
}
\description{
This function splits groups of polygons grouped by date proximity into smaller groups to ensure that the max number of polygons per group is respected.
}
