% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{tmin_tmax_interpolation}
\alias{tmin_tmax_interpolation}
\title{Interpolate Daily Minimum and Maximum Temperatures to Obtain Hourly Values}
\usage{
tmin_tmax_interpolation(
  paths,
  polygons,
  boundary_dates,
  interpol_fun,
  interpol_args = NULL,
  daily_agg_fun = "none",
  max_cells = 3e+07,
  max_output_layers = 30000,
  save_interpol_rast = FALSE
)
}
\arguments{
\item{paths}{A list containing paths to the input temperature raster files, including
paths for daily minimum and maximum temperatures.}

\item{polygons}{An object that defines geographical boundaries, used for cropping the rasters.}

\item{boundary_dates}{A vector of dates defining the range for which interpolation
is to be performed.}

\item{interpol_fun}{A function used for interpolating the temperatures. This function
should accept minimum and maximum temperatures as inputs.}

\item{interpol_args}{A list of additional arguments to be passed to the interpolation function.}

\item{daily_agg_fun}{A character string specifying the aggregation function to be used
when converting back to daily values. Defaults to "none".}

\item{max_output_layers}{An integer defining the maximum number of output layers allowed;
default is set to 30,000 to prevent exceeding raster writing limits.}
}
\value{
A SpatRaster object containing the daily aggregated temperatures calculated from the
interpolated hourly values based on the provided minimum and maximum temperatures.
}
\description{
This function interpolates daily minimum and maximum temperature raster values into
hourly temperature values using a specified interpolation function. The function
also aggregates the hourly temperature data back to daily values if requested.
It handles large raster datasets by splitting them into manageable batches to avoid
processing limitations.
}
\examples{
  # Example using the tmin_tmax_interpolation function:
  result <- tmin_tmax_interpolation(
    paths = list(path_tmin_rast = "path/to/tmin",
                 path_tmax_rast = "path/to/tmax"),
    polygons = my_polygons,
    boundary_dates = as.Date(c("2022-01-01", "2022-01-07")),
    interpol_fun = sinusoidal_interpolation,
    interpol_args = list(),
    daily_agg_fun = "mean"
  )
}
