% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{transform_binned_output}
\alias{transform_binned_output}
\title{Transform Binned Output}
\usage{
transform_binned_output(data, temporal_resolution)
}
\description{
Transform Binned Output
}
\keyword{internal}
