% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_validation.R
\name{validate_errors}
\alias{validate_errors}
\title{Access and make the error list more accesible}
\usage{
validate_errors(errors_list)
}
\arguments{
\item{errors_list}{List of errors returned by the shocks_wrapper function}
}
\value{
Null if no errors exist. Else, a list with the different types of errors encountered.
}
\description{
This function accesses the error list returned by the shocks_wrapper function and creates a tidier error list
}
