% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_validation.R
\name{validate_global_summaries}
\alias{validate_global_summaries}
\title{Get basic summaries and NA counts for the shocks measures}
\usage{
validate_global_summaries(results_data)
}
\arguments{
\item{results_data}{A dataframe output by the shocks_wrapper function}
}
\value{
Null. Prints the summary for each measure and the NA count in the console.
}
\description{
This function print basic statistical summaries and NA counts for each measure in the dataframe shocks information
}
