% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Function to validate inputs}
\usage{
validate_inputs(
  pol_date_pairs,
  conditions_list,
  window,
  start,
  hist_lags,
  align,
  bin_width,
  prop_cores,
  int_threshold
)
}
\arguments{
\item{pol_date_pairs}{A data.frame with polygons and dates}

\item{conditions_list}{A named list with conditions to query the parquet file.}

\item{window}{An integer indicating wht length of the date sequence}

\item{start}{An integer indicating the offset with respect to the polygon date.}

\item{hist_lags}{An integer indicating the number of years used to build the dynamic historical window.}

\item{align}{A character string indicating whether the sequence is left-, or right-aligned with respect to the polygon date.}

\item{bin_width}{An integer indicating the size of the bins into which the date sequence is divided.}

\item{prop_cores}{A float number, between 0 and 1, indicating the proportion od available cores to use.}

\item{int_threshold}{A float number, between 0 and 1, indicating the minimum intersection threshold with the parquet dates needed for a historic sequence to be valid.}
}
\value{
TRUE if all inputs are valid; otherwise raises an error.
}
\description{
This functions validates that the inputs passed to the shocks_wrapper function are correct.
}
