% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_validation.R
\name{validate_measure_distributions}
\alias{validate_measure_distributions}
\title{Produce plots to visualize the distibution of each measure on the dataframe returned by the shocks_wrapper function}
\usage{
validate_measure_distributions(results_data, show_plots = TRUE)
}
\arguments{
\item{results_data}{A dataframe returned by the shocks wrapper function}

\item{show_plots}{A boolean indicating whether to print the plots in the plot console}
}
\value{
A list with plots for each measure
}
\description{
This function plots the measures in the dataframe returned by the shocks wrapper function
}
