% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{window_seq}
\alias{window_seq}
\title{Given a baseline date, return a sequence of dates with a specified alignment, width and lags.}
\usage{
window_seq(
  d,
  start,
  window,
  align,
  hist_lags,
  window_spec,
  start_date,
  stop_date,
  only_hist = FALSE
)
}
\arguments{
\item{d}{A character with dates in format YYYY-MM-DD}

\item{start}{An integer indicating the offset with respect to the polygon date.}

\item{window}{An integer indicating the size of the date sequences.}

\item{align}{A character indicating whether the window sequence is left- or right-aligned with respect to the polygon date.}

\item{hist_lags}{An integer indicating the number of historical years in the dynamic baseline.}

\item{window_spec}{A character vector indicating whether the historic window is dynamic, fixed or both.}

\item{start_date}{A date in the format YYYY-MM-DD indicating the start date of the fixed window. Only required if window_spec = "fixed" or window_spec = "both".}

\item{stop_date}{A date in the format YYYY-MM-DD indicating the end date of the fixed window. Only required if window_spec = "fixed" or window_spec = "both".}

\item{only_hist}{A boolean indicating whether to return only the historical sequences (if TRUE) or not (if FALSE). Default is FALSE.}
}
\value{
A vector of dates
}
\description{
This function takes a baseline date an returns and sequence of historic dates.
}
