% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uval.R
\name{calculateUval}
\alias{calculateUval}
\title{calculateUval}
\source{
DIN 4108-3
}
\usage{
calculateUval(Layers, k.t_e = -10, k.t_i = 20, k.h_e = 25, k.h_i = 8,
  k.rf_e = 80, k.rf_i = 50, k.p_e = 1168, k.p_i = 321,
  plottemp = FALSE, plotpress = FALSE, writetab = FALSE, name = "")
}
\arguments{
\item{Layers,}{building elements layers}

\item{k.t_e}{(optional, default = -10) External air temperature [C]}

\item{k.t_i}{(optional, default = 20) Internal air temperature [C]}

\item{k.h_e}{(optional, default = 25) External heat transfer coefficient
[m2K/W]}

\item{k.h_i}{(optional, default = 8) Internal heat transfer coefficient
[m2K/W]}

\item{k.rf_e}{(optional, default = 80) External air moisture}

\item{k.rf_i}{(optional, default  = 50) Internal air moisture}

\item{k.p_e}{(optional, default = 1168) External pressure [Pa]}

\item{k.p_i}{(optional, default = 321) Internal pressure [Pa]}

\item{plottemp}{(optional, default = FALSE) Store a temperature profile plot
under "./FIGURES/[name]_pressure.pdf" and
"./FIGURES/[name]_temperature.pdf"}

\item{plotpress}{(optional, default = FALSE) Store a pressure profile plot
under "./FIGURES"}

\item{writetab}{(optional, default = FALSE) Store three tables for pressure,
temperature and material properties as "./TABLES/[name]_PressProfile.csv",
"./TABLES/[name]_TempProfile.csv" and "./TABLES/[name]_Layers.csv"}

\item{name}{(optional = "")}
}
\value{
Uval component U-value
}
\description{
U-value calculator
}
\examples{
# Reproduces the example 4 from DIN 4108-3
MaterialsToget <- c(
  "Mineralischer Edelputz",
  "Porensinterbeton mit Quarzsand_900",
  "EPS_040.30",
  "Holzwolle-Leichtbauplatten_Heraklith-Platten (Magnesia)_390",
  "Mineralischer Armierungsputz"
  )

Thicknes = c(0.02,0.24,0.04,0.025,0.015)

# Generate the Layers
Layers <- getLayers(MaterialsToget, Thicknes)

uval <- calculateUval(Layers, k.t_e=-5,
                      plottemp=TRUE, plotpress=TRUE,
                      name="ExternalData_", writetab=TRUE)

# from DIN 4108-3
# B.5 Beispiel 4: Außenwand mit nachträglicher raumseitiger Wärmedämmung
# Konstruktion: Außenwand mit nachträglicher Innendämmung
Layers <- data.frame(
    Layer = c(
        "Außenputz",
        "Mauerwerk",
        "EPS-Dämmstoff",
        "HWL-Platte",
        "Innenputz"),
    Thicknes = c(0.02,0.24,0.04,0.025,0.015),
    Conductivity = c(1,0.4,0.04,0.08,0.7),
    Diffusion = c(40,8,20,4,15))

uval <- calculateUval(Layers, k.t_e=-5,
                      plottemp=TRUE, plotpress=TRUE,
                      name="Innendaemmung_", writetab=TRUE)

# from DIN 4108-3
# Konstruktion: Leichte Außenwand mit hinterlüfteter Vorsatzschale
Layers <- data.frame(
    Layer = c(
        "Vorgehängte Außenschale",
        "Belüftete Luftschicht",
        "Spanplatte V100",
        "Mineralwolle",
        "Diffusionshemmende Schicht",
        "Spanplatte V20"),
    Thicknes = c(0.02, 0.03, 0.019, 0.16, 0.00005, 0.019),
    Conductivity = c(NaN, NaN, 0.127, 0.04, NaN, 0.127),
    Diffusion = c(NaN, NaN, 100, 1, 40000, 50))

uval <- calculateUval(Layers, k.t_e=-5,
                      plottemp=TRUE, plotpress=TRUE,
                      name="leichteAussenwand_", writetab=TRUE)

# from Keller, B., & Rutz, S. (2010).
# Pinpoint: Key facts + figures for sustainable buildings. Basel: Birkhauser.
Layers <- data.frame(
    Layer = c("External Plaster", "Expanded Polystyrene",
              "Reinforced Concrete", "Internal Stuco"),
    Thicknes =     c(0.01, 0.18 , 0.25, 0.01),
    Conductivity = c(0.87, 0.038, 1.8,  0.7),
    Diffusion =    c(15, 30, 80, 6))

uval <- calculateUval(Layers, k.t_e=-5,
                      plottemp=TRUE, plotpress=TRUE,
                      name="ReinforcedConcrete_", writetab=TRUE)
}
\author{
M. Estebna Munoz H.
}

