% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shocks_utils.R
\name{create_bin_scheme}
\alias{create_bin_scheme}
\title{Create bin structure}
\usage{
create_bin_scheme(window, bin_width, align, disjoint = FALSE)
}
\arguments{
\item{window}{Integer specifying the size of the sequence of dates.}

\item{bin_width}{Integer indicating the size of each bin. Within the window. If window \%\% bin_width == 0, then n bins of size bin_width and 1 bin of size y (the remainder) are returned such that window = n * bin_width + y. Otherwise, n bins of size bin_Width are returned.}

\item{align}{Character indicating the alignment of the window with respect to the interview date.}

\item{disjoint}{whether the bins are disjoint or not. Default is to return overlapping bins.}
}
\value{
A named list with boolean vectors representing the bin scheme.
}
\description{
This function creates a list of boolean vectors that indicate the position of the dates that fall into the specified bins.
}
\examples{
:
\dontrun{
  create_bin_scheme(10, 5, "right", disjoint = FALSE)
}
}
