% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data.R
\name{filter_env_rast}
\alias{filter_env_rast}
\title{Filter Raster by Date Range}
\usage{
filter_env_rast(env_rast, start_date, end_date, verbose = 1)
}
\arguments{
\item{:}{\itemize{
\item env_rast: A multi-layer raster object (e.g., from the 'raster' or 'terra' package) with layer names
corresponding to valid date strings (e.g., 2017-01-01).
\itemize{
\item start_date: A string or Date object representing the start date of the desired date range.
\item end_date: A string or Date object representing the end date of the desired date range.
}
}}
}
\value{
:
\itemize{
\item A filtered raster object containing only the layers corresponding to the specified date range.
}
}
\description{
This function filters a multi-layer raster based on a specified date range. The raster should have
layer names corresponding to date strings in 'YYYY-MM-DD' format. The function checks whether the
specified date range falls within the range of available raster dates and returns the filtered
raster containing only the layers that match the specified dates.
}
\examples{
  # Assuming 'my_raster' is a multi-layer raster and you wish to filter
  # it for dates between January 1, 2018, and December 31, 2018:
  filtered_raster <- filter_env_rast(my_raster, start_date = "2018-01-01", end_date = "2018-12-31")
}
