% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\docType{package}
\name{heat-package}
\alias{heat}
\alias{heat-package}
\title{heat: WHO HEAT Application}
\description{
An application to explore and compare within-country health 
   inequalities across settings.
}
\author{
\strong{Maintainer}: Zev Ross \email{zev@zevross.com}

Authors:
\itemize{
  \item Ahmadreza Hosseinpoor [contributor]
  \item Anne Schlotheuber [contributor]
}

Other contributors:
\itemize{
  \item Nate Teetor \email{nate@zevross.com} [contributor]
  \item  World Health Organization [copyright holder]
}

}
