% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heat.R
\name{heat.default}
\alias{heat.default}
\title{heat}
\source{
DIN V 18599
}
\usage{
\method{heat}{default}(building = FALSE, climate = FALSE,
  general_name = "No Name", general_ploto = FALSE,
  general_plotn = "No Name", building_uwb = FALSE,
  building_windows = FALSE, building_uvalw = FALSE,
  building_uvalr = FALSE, building_uvalwindow = FALSE,
  building_dim = c(FALSE, FALSE), building_h = FALSE,
  building_aircrate = FALSE, building_ti = FALSE, building_qi = FALSE,
  building_orientation = FALSE, building_storagecapacity = FALSE,
  building_roofslope = FALSE, output_type = "Year")
}
\arguments{
\item{building}{(optional, default = FALSE) use a user define building
data file.}

\item{climate}{(optional, default = FALSE) use a specific climate file.}

\item{general_name}{(optional, default = "No Name") if this value is not
specifically define the function will load the values from the building data
file.}

\item{general_ploto}{(optional, default = FALSE)}

\item{general_plotn}{(optional, default="No Name")}

\item{building_uwb}{(optional, default = FALSE)}

\item{building_windows}{(optional, default = FALSE)}

\item{building_uvalw}{(optional, default = FALSE)}

\item{building_uvalr}{(optional, default = FALSE)}

\item{building_uvalwindow}{(optional, default = FALSE)}

\item{building_dim}{(optional, default = c(FALSE,FALSE))}

\item{building_h}{(optional, default = FALSE)}

\item{building_aircrate}{(optional, default = FALSE)}

\item{building_ti}{(optional, default = FALSE)}

\item{building_qi}{(optional, default = FALSE)}

\item{building_orientation}{(optional, default = FALSE)}

\item{building_storagecapacity}{(optional, default = FALSE)}

\item{building_roofslope}{(optional, default = FALSE)}

\item{output_type}{(optional, default = "Year")}
}
\description{
Perform a monthly heat balance and annual energy demand of buildings
}
\examples{
# Use the default values to compute the heat demand of a building
climate_month_names <- c("January","February","March","April",
                         "May","June","July","August",
                         "September","October","November","December")

model <- heat(building_orientation = 0,
              output_type = "Month")
Qhm <- model$Qhm
barplot(Qhm,
        names.arg=climate_month_names,
        main = "Monthly Heat Demand",
        ylab =  "head Demand [kWh]")

# This example show the heat demand variation for different
# U values combinations
Buildings.Number <- 9

U.Values <- matrix(c(
    1.3,    1.0,	3.0,    # 01
    1.2,	0.9,	2.7,    # 02
    1.1,	0.8,	2.7,    # 03
    1.0,	0.7,	2.7,    # 04
    0.9,	0.6,	2.4,    # 05
    0.8,	0.5,	2.1,    # 06
    0.6,	0.4,	1.9,    # 07
    0.5,	0.3,	1.6,    # 08
    0.4,	0.2,	1.6),   # 09
    3,Buildings.Number)

Heat.Demand <- rep(0,Buildings.Number)
for (i in 1:Buildings.Number){
    UvalW <- U.Values[1,i]
    UvalR <- U.Values[2,i]
    UvalWindow <- U.Values[3,i]
    temp.2 <- heat(building_uvalw = UvalW,
                    building_uvalr = UvalR,
                    building_uvalwindow = UvalWindow)
    Heat.Demand[i] <- temp.2$Qhs
}

# bar plot
barplot(Heat.Demand,
        names.arg = seq(1910,2010,12),
        main = "Heat demand for a set of buildings",
        ylab = "Heat demand [kWh/m²a]",
        xlab = "Building age")

# This example show the heat demand variation for different
# user influenced parameters
Buildings.Number <- 5

Param.Values <- matrix(c(
    7, 22, 0.7,   # 01
    6, 21, 0.6,   # 02
    5, 20, 0.5,   # 03
    4, 19, 0.4,   # 04
    3, 18, 0.3),  # 05
    3,Buildings.Number)

UvalW <- 0.4
UvalR <- 0.2
UvalWindow <- 1.6

Heat.Demand <- rep(0,Buildings.Number)
for (i in 1:Buildings.Number){
    AirCRate <- Param.Values[1,i]
    Ti <- Param.Values[2,i]
    qi <- Param.Values[3,i]
    temp.2 <- heat(building_aircrate = AirCRate,
                   building_ti = Ti,
                   building_qi = qi,
                   building_uvalw = UvalW,
                   building_uvalr = UvalR,
                   building_uvalwindow = UvalWindow)
    Heat.Demand[i] <- temp.2$Qhs
}

# bar plot
barplot(Heat.Demand,
        main = "Heat demand for a set of buildings",
        ylab = "Heat demand [kWh/m²a]",
        xlab = "Building age")

# This example show the heat demand variation for different
# Building orientations, it used the ggplot2 library
library(ggplot2)

iter <- seq(0,360,1)
Heat.Demand = rep(0,length(iter))
Heat.Gains.Solar = rep(0,length(iter))
Irradiation.Sum = rep(0,length(iter))

building_dim <- c(12,6)
for (i in 1:length(iter)){
    BO <- iter[i]
    temp.2 <- heat(building_orientation = BO,
                   building_dim = building_dim)
    Heat.Demand[i] <- temp.2$Qhs
    Heat.Gains.Solar[i] <- temp.2$Ss
    Irradiation.Sum[i] <- temp.2$Ti
}

# Polar plot + line plot
result <- data.frame(heat.demand = Heat.Demand,
                     orientation = iter)
doh <- ggplot(result, aes(orientation, heat.demand))
# Line plot
doh + geom_line(colour = "red", size = 1)  +
    coord_polar(direction = -1, start = -pi/2) +
    labs(title = "Heat demand for all possible building orientations") +
    scale_x_continuous(breaks=seq(0, 360, 15))
}
\author{
M. Esteban Munoz H.
}

