% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpol_min_max}
\alias{interpol_min_max}
\title{Interpolate Daily Minimum and Maximum Temperatures to Obtain Hourly Values}
\usage{
interpol_min_max(
  min_rast_path,
  max_rast_path,
  geometry,
  start_date,
  end_date,
  interpol_fun,
  interpol_args = NULL,
  daily_agg_fun = "none",
  save_path = NULL,
  max_cells = 3e+07,
  max_output_layers = 30000
)
}
\arguments{
\item{min_rast_path}{Path to the directory containing minimum temperature raster files.}

\item{max_rast_path}{Path to the directory containing maximum temperature raster files.}

\item{geometry}{An sf object that defines geographical boundaries, used for cropping the rasters.}

\item{start_date}{Character string or Date object specifying the start date for interpolation (e.g., "2022-01-01").}

\item{end_date}{Character string or Date object specifying the end date for interpolation (e.g., "2022-01-07").}

\item{interpol_fun}{A function used for interpolating the temperatures. This function
should accept minimum and maximum temperatures as inputs.}

\item{interpol_args}{A list of additional arguments to be passed to the interpolation function.}

\item{daily_agg_fun}{A character string specifying the aggregation function to be used
when converting back to daily values. Defaults to "none".}

\item{save_path}{Optional character string specifying the output directory for storing
the interpolated raster. If NULL (default), the raster is only returned
without saving to disk. When provided, the output raster is saved as a
GeoTIFF split up into batches.}

\item{max_output_layers}{An integer defining the maximum number of output layers allowed;
default is set to 30,000 to prevent exceeding raster writing limits.}
}
\value{
A SpatRaster object containing the daily aggregated temperatures calculated from the
interpolated hourly values based on the provided minimum and maximum temperatures.
}
\description{
This function interpolates daily minimum and maximum raster values using a specified
interpolation function. The function
also aggregates the hourly data back to daily values if requested.
It handles large raster datasets by splitting them into manageable batches to avoid
processing limitations.
}
\examples{
\dontrun{
  # Example using the interpol_min_max function:
  result <- interpol_min_max(
    min_rast_path = "path/to/tmin",
    max_rast_path = "path/to/tmax",
    geometry = my_regions,
    start_date = "2022-01-01",
    end_date = "2022-01-07",
    interpol_fun = sinusoidal_interpol,
    interpol_args = list(),
    daily_agg_fun = "mean"
  )
}

}
